// ecommerce-application/src/widgets/RegistrationForm/lib/types/RegistrationFields.ts
export interface IRegistrationFields {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  dateOfBirth: Date;
  shippingStreet: string;
  shippingCity: string;
  shippingPostalCode: string;
  shippingCountry: string;
  billingStreet: string;
  billingCity: string;
  billingPostalCode: string;
  billingCountry: string;
}